games/dino: drop redundant zero-inits and shrink config tables
The runner/trex/distanceMeter/horizon/horizonLine globals live in BSS,
and their init functions run exactly once, so writing zeros there is
dead weight; the restart path sets its own state and is untouched.
Config tables shrink to short (largest value is 999), the system color
table is gone - the window background is the game background color,
visible only in the few border pixels around the back buffer. Prune
dead includes while at it.

.text 12317 -> 11947, .data -136. The 20000-frame horizon simulation
stays byte-identical.

+++ programs/games/dino/obstacle.c
@@ -59,8 +59,8 @@ ObstacleTypeConfig obstacleTypeConfigs[3] = {
 		}
 };
 
-int obstacleSpritePosX[3] = { ATLAS_CACTUS_SMALL_X, ATLAS_CACTUS_LARGE_X, ATLAS_PTERODACTYL_X};
-int obstacleSpritePosY[3] = { ATLAS_CACTUS_SMALL_Y, ATLAS_CACTUS_LARGE_Y, ATLAS_PTERODACTYL_Y};
+short obstacleSpritePosX[3] = { ATLAS_CACTUS_SMALL_X, ATLAS_CACTUS_LARGE_X, ATLAS_PTERODACTYL_X};
+short obstacleSpritePosY[3] = { ATLAS_CACTUS_SMALL_Y, ATLAS_CACTUS_LARGE_Y, ATLAS_PTERODACTYL_Y};
 
 
 void obstacleInit(Obstacle* ob, const ObstacleTypeConfig *otc, double speed, int opt_xOffset) {
